+++ cpp/GlobalObject.h
@@ -44,7 +44,8 @@ public:
     Q_INVOKABLE void setStatusMessage(QString msg) const;
     Q_INVOKABLE QString getApplicationVersion(){return QString::fromStdString(VERSION_APPLICATION);}
     Q_INVOKABLE QString checkMatrixServerUrl(QString url);
-
+    Q_INVOKABLE QString materialPrimaryColor();
+    
 public slots:
     bool handleMatrixUri(const QByteArray &uri);
     bool handleMatrixUri(const QUrl &uri);